start to move playback stuff back into session, and re-do saved tests table
