+++ app/renderer/components/Session/Playback.css
