+++ app/renderer/components/Session/Session.js
@@ -3,20 +3,28 @@ import React, { Component } from 'react';
 import NewSessionForm from './NewSessionForm';
 import SavedSessions from './SavedSessions';
 import AttachToSession from './AttachToSession';
+import SessionModeTabs from './SessionModeTabs';
 import ServerTypeTabs from './ServerTypeTabs';
+import SavedTests from './SavedTests';
 import { Tabs, Button, Spin, Icon, Card } from 'antd';
 import SessionStyles from './Session.css';
+import { SessionModes } from '../../actions/Session';
 
 const {TabPane} = Tabs;
 
 export default class Session extends Component {
 
   componentWillMount () {
-    const {setLocalServerParams, getSavedSessions, setSavedServerParams, getRunningSessions} = this.props;
+    const {
+      setLocalServerParams, getSavedSessions, setSavedServerParams,
+      getRunningSessions, getSavedTests, getTestResults,
+    } = this.props;
     (async function () {
       await getSavedSessions();
       await setSavedServerParams();
       await setLocalServerParams();
+      await getSavedTests();
+      await getTestResults();
       getRunningSessions();
     })();
   }
@@ -24,33 +32,30 @@ export default class Session extends Component {
   render () {
     const {newSessionBegan, savedSessions, tabKey, switchTabs,
       requestSaveAsModal, newSession, caps, capsUUID, saveSession, isCapsDirty,
-      sessionLoading, attachSessId} = this.props;
-
-    let mode = "inspect";
+      sessionLoading, attachSessId, mode} = this.props;
 
     const isAttaching = tabKey === 'attach';
+    const showCapsTab = !newSessionBegan && mode !== SessionModes.view;
+    const showSavedTests = mode === SessionModes.playback;
 
     return <Spin spinning={!!sessionLoading}>
       <div className={SessionStyles['session-container']}>
-        <Tabs activeKey={mode} className={SessionStyles.topmostTabs}>
-          <TabPane tab='Inspect & Record Tests' key="inspect">
-          </TabPane>
-          <TabPane tab='Run a Saved Test' key="playback">
-            <div>Choose a previously-recorded test to run.</div>
-          </TabPane>
-          <TabPane tab='View Saved Tests' key="view">
-          </TabPane>
-        </Tabs>
-        <Card className={SessionStyles.sessionCard}>
+        <SessionModeTabs {...this.props} />
+
+        {showSavedTests && <Card className={`${SessionStyles.sessionCard} ${SessionStyles.sessionCardTests}`}>
+            <SavedTests {...this.props} />
+        </Card>}
+
+        {showCapsTab && <Card className={SessionStyles.sessionCard}>
           <ServerTypeTabs {...this.props} />
-        </Card>
+        </Card>}
 
 
         {newSessionBegan && <div key={2}>
           <p>Session In Progress</p>
         </div>}
 
-        {!newSessionBegan && <Card className={`${SessionStyles.sessionCard} ${SessionStyles.sessionCardCaps}`}>
+        {showCapsTab && <Card className={`${SessionStyles.sessionCard} ${SessionStyles.sessionCardCaps}`}>
           <Tabs activeKey={tabKey} onChange={switchTabs} className={SessionStyles.scrollingTabCont}>
             <TabPane tab='Desired Capabilities' key='new' className={SessionStyles.scrollingTab}>
               <NewSessionForm {...this.props} />
@@ -63,6 +68,8 @@ export default class Session extends Component {
             </TabPane>
           </Tabs>
         </Card>}
+
+        {showCapsTab &&
         <div className={SessionStyles.sessionFooter}>
           <div style={{float: 'left'}}>
             <a href="#" onClick={(e) => e.preventDefault() || shell.openExternal("https://github.com/appium/appium/blob/master/docs/en/writing-running-appium/caps.md")}>
@@ -80,7 +87,7 @@ export default class Session extends Component {
               Attach to Session
             </Button>
           }
-        </div>
+        </div>}
       </div>
     </Spin>;
   }
