+++ app/renderer/routes.js
@@ -5,7 +5,6 @@ import StartServerPage from './containers/StartServerPage';
 import ServerMonitorPage from './containers/ServerMonitorPage';
 import SessionPage from './containers/SessionPage';
 import InspectorPage from './containers/InspectorPage';
-import PlaybackLibraryPage from './containers/PlaybackLibraryPage';
 import UpdaterPage from './containers/UpdaterPage';
 
 export default (
@@ -13,7 +12,6 @@ export default (
     <IndexRoute component={StartServerPage} />
     <Route path="monitor" component={ServerMonitorPage} />
     <Route path="session" component={SessionPage} />
-    <Route path="playback" component={PlaybackLibraryPage} />
     <Route path="inspector" component={InspectorPage} />
     <Route path="updater" component={UpdaterPage} />
   </Route>
